Split services features into multiple for individual services
